third_party/WebKit/LayoutTests/inspector/debugger/source-url-comment-expected.txt

Tests that evals with sourceURL comment are shown in scripts panel.

Debugger was enabled.

Running: testSourceURLCommentInInlineScript

Running: testSourceURLComment
function keepAlive() {}

Running: testDeprecatedSourceURLComment
function keepAlive() {}

Running: testSourceURLCommentInDynamicScript
function keepAliveInDynamicScript() {}

Running: testNonRelativeURL
function relativeURLScript() {}
Debugger was disabled.
